requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
certifi==2019.9.11
chardet==3.0.4
Django==2.2.5
django-filter==2.2.0
djangorestframework==3.10.3
idna==2.8
importlib-metadata==0.23
Markdown==3.1.1
more-itertools==7.2.0
packaging==19.2
pluggy==0.13.0
py==1.8.0
pyparsing==2.4.2
pytz==2019.2
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.6
wcwidth==0.1.7
zipp==0.6.0
